+++ R/metaprop.R
@@ -790,8 +790,8 @@ metaprop <- function(event, n, studlab,
                      ##
                      incr = gs("incr"), allincr = gs("allincr"),
                      addincr = gs("addincr"),
-                     method.ci = gs("method.ci"),
                      ##
+                     method.ci = gs("method.ci.prop"),
                      level = gs("level"), level.comb = gs("level.comb"),
                      comb.fixed = gs("comb.fixed"),
                      comb.random = gs("comb.random"),